pyproject.toml
[tool.poetry]
name = "wifiqr"
version = "0.1.0"
description = "Generate JPG WiFi QR codes from environment variables"
authors = ["Teo Sibileau <teo@example.com>"]
readme = "README.md"
packages = [
    { include = "wifiqr", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.10"
qrcode = "^8.2"
python-dotenv = "^1.2.1"
click = "^8.3.0"
Pillow = "^10.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"

[tool.poetry.scripts]
wifiqr = "wifiqr.generator:main"

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "DTZ", "T10", "EM", "ISC", "ICN", "G", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SIM", "TID", "TCH", "ARG", "PTH", "ERA", "PGH", "PL", "TRY", "FLY", "NPY", "RUF"]
ignore = ["E501", "PLR0913", "ARG002"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.bandit]
exclude_dirs = ["tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
